feat: send requests to userserver for auth
This adds functionality for sending requests to the REST
API and handling the responses, and uses this functionality
during registration and login to authenticate via the server.

Note that login currently doesn't work (or rather always
returns and subsequently ignores 405) as the server is
not set up to accept POSTs on /login.

This also removes the unused layout/code for validating
the user's email address. This can be re-added later
if we decide to implement this functionality in the server.